refactor(lib/core/math): move rand into it's own module, should improve bundle sizes where it's not
