+++ src/lab/04-vectors/src/main.ts
@@ -1,6 +1,7 @@
 import { el } from '../../../lib/core/dom';
 import loop from '../../../lib/game/loop';
-import { random, sin, cos, ππ, round } from '../../../lib/core/math';
+import { sin, cos, ππ, round } from '../../../lib/core/math';
+import { random } from '../../../lib/core/rand';
 import resize from './resize';
 import { vec2 } from '../../../lib/vec';
 import { sub, mul, add, div } from '../../../lib/vec/math';
