+++ src/lib/core/math.test.ts
@@ -1,44 +1,10 @@
-import {
-  lerp,
-  randomBool,
-  randomInt,
-  randomRange,
-  saw,
-  toDegrees,
-  toRadians,
-  tri,
-  π,
-  ππ,
-} from './math';
+import { lerp, saw, toDegrees, toRadians, tri, π, ππ } from './math';
 
 describe('@fartts/lib/math', () => {
   test('lerp', () => {
     expect(lerp(1, 2, 0.5)).toBe(1.5);
     expect(lerp(0, 2, 0.5)).toBe(1);
     expect(lerp(1, 5, 0.25)).toBe(2);
-  });
-
-  test('randomBool', () => {
-    expect(randomBool()).toBe(false);
-    expect(randomBool()).toBe(false);
-    expect(randomBool()).toBe(false);
-    expect(randomBool()).toBe(false);
-    expect(randomBool()).toBe(true);
-  });
-
-  test('randomRange', () => {
-    expect(randomRange(0, 1)).toBe(0.3858736207545881);
-    expect(randomRange(1, 2)).toBe(1.2395959625293629);
-    expect(randomRange(1, 30)).toBe(27.889704444634948);
-    expect(randomRange(40, 20)).toBe(24.1776263921708);
-    expect(randomRange(5000, 5)).toBe(3343.64129004656);
-  });
-
-  test('randomInt', () => {
-    expect(randomInt(10)).toBe(3);
-    expect(randomInt(-20)).toBe(-11);
-    expect(randomInt(40, 20)).toBe(32);
-    expect(randomInt(5, 5000)).toBe(1675);
   });
 
   test('saw', () => {
