+++ src/lib/core/math.ts
@@ -1,7 +1,3 @@
-import seedRandom from 'seedrandom';
-const name = '@fartts/fartts';
-export const random = seedRandom(name);
-
 export const {
   abs,
   acos,
@@ -41,40 +37,6 @@ export function lerp(v0: number, v1: number, t: number): number {
    * @see: https://en.wikipedia.org/wiki/Linear_interpolation#Programming_language_support
    */
   return (1 - t) * v0 + t * v1;
-}
-
-/**
- * ## randomBool
- *
- * @export
- * @returns {boolean} a random boolean value
- */
-export function randomBool(): boolean {
-  return round(random()) === 1;
-}
-
-/**
- * ## randomInt
- *
- * @export
- * @param {number} a
- * @param {number} [b]
- * @returns {number}
- */
-export function randomInt(a: number, b?: number): number {
-  return round(b === undefined ? randomRange(0, a) : randomRange(a, b));
-}
-
-/**
- * ## randomRange
- *
- * @export
- * @param {number} a
- * @param {number} b
- * @returns {number}
- */
-export function randomRange(a: number, b: number): number {
-  return lerp(a, b, random());
 }
 
 /**
